feat: add useWallets hook
