+++ packages/widget/src/pages/SelectWalletPage/SelectWalletPage.tsx
@@ -1,55 +1,33 @@
-import { ChainType } from '@lifi/sdk';
-import type { CreateConnectorFnExtended } from '@lifi/wallet-management';
-import {
-  createCoinbaseConnector,
-  createMetaMaskConnector,
-  createWalletConnectConnector,
-  isWalletInstalled,
-} from '@lifi/wallet-management';
-import type { Theme } from '@mui/material';
 import {
   Button,
   DialogActions,
   DialogContent,
   DialogContentText,
   List,
-  useMediaQuery,
 } from '@mui/material';
-import { WalletReadyState } from '@solana/wallet-adapter-base';
 import type { Wallet } from '@solana/wallet-adapter-react';
-import { useWallet } from '@solana/wallet-adapter-react';
-import { useCallback, useMemo, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { Connector } from 'wagmi';
-import { useConnect, useAccount as useWagmiAccount } from 'wagmi';
+import { useAccount as useWagmiAccount } from 'wagmi';
 import { Dialog } from '../../components/Dialog.js';
 import { PageContainer } from '../../components/PageContainer.js';
-import { defaultCoinbaseConfig } from '../../config/coinbase.js';
-import { defaultMetaMaskConfig } from '../../config/metaMask.js';
-import { defaultWalletConnectConfig } from '../../config/walletConnect.js';
 import { useHeader } from '../../hooks/useHeader.js';
+import { useWallets } from '../../hooks/useWallets.js';
 import { useWidgetConfig } from '../../providers/WidgetProvider/WidgetProvider.js';
-import { isItemAllowed } from '../../utils/item.js';
 import { EVMListItemButton } from './EVMListItemButton.js';
 import { SVMListItemButton } from './SVMListItemButton.js';
-import { walletComparator } from './utils.js';
 
 export const SelectWalletPage = () => {
   const { t } = useTranslation();
   const { chains, walletConfig } = useWidgetConfig();
   const account = useWagmiAccount();
-  const { connectors } = useConnect();
   const [walletIdentity, setWalletIdentity] = useState<{
     show: boolean;
     connector?: Connector;
   }>({ show: false });
-  const { wallets: solanaWallets } = useWallet();
 
   useHeader(t(`header.selectWallet`));
-
-  const isDesktopView = useMediaQuery((theme: Theme) =>
-    theme.breakpoints.up('sm'),
-  );
 
   const closeDialog = () => {
     setWalletIdentity((state) => ({
@@ -65,92 +43,7 @@ export const SelectWalletPage = () => {
     });
   }, []);
 
-  const wallets = useMemo(() => {
-    const evmConnectors: (CreateConnectorFnExtended | Connector)[] =
-      Array.from(connectors);
-    if (
-      !connectors.some((connector) =>
-        connector.id.toLowerCase().includes('walletconnect'),
-      )
-    ) {
-      evmConnectors.unshift(
-        createWalletConnectConnector(
-          walletConfig?.walletConnect ?? defaultWalletConnectConfig,
-        ),
-      );
-    }
-    if (
-      !connectors.some((connector) =>
-        connector.id.toLowerCase().includes('coinbase'),
-      ) &&
-      !isWalletInstalled('coinbase')
-    ) {
-      evmConnectors.unshift(
-        createCoinbaseConnector(
-          walletConfig?.coinbase ?? defaultCoinbaseConfig,
-        ),
-      );
-    }
-    if (
-      !connectors.some((connector) =>
-        connector.id.toLowerCase().includes('metamask'),
-      ) &&
-      !isWalletInstalled('metaMask')
-    ) {
-      evmConnectors.unshift(
-        createMetaMaskConnector(
-          walletConfig?.metaMask ?? defaultMetaMaskConfig,
-        ),
-      );
-    }
-    const evmInstalled = isItemAllowed(ChainType.EVM, chains?.types)
-      ? evmConnectors.filter(
-          (connector) =>
-            isWalletInstalled(connector.id!) &&
-            // We should not show already connected connectors
-            account.connector?.id !== connector.id,
-        )
-      : [];
-    const evmNotDetected = isItemAllowed(ChainType.EVM, chains?.types)
-      ? evmConnectors.filter((connector) => !isWalletInstalled(connector.id!))
-      : [];
-    const svmInstalled = isItemAllowed(ChainType.SVM, chains?.types)
-      ? solanaWallets?.filter(
-          (connector) =>
-            connector.adapter.readyState === WalletReadyState.Installed &&
-            // We should not show already connected connectors
-            !connector.adapter.connected,
-        )
-      : [];
-    const svmNotDetected = isItemAllowed(ChainType.SVM, chains?.types)
-      ? solanaWallets?.filter(
-          (connector) =>
-            connector.adapter.readyState !== WalletReadyState.Installed,
-        )
-      : [];
-
-    const installedWallets = [...evmInstalled, ...svmInstalled].sort(
-      walletComparator,
-    );
-
-    if (isDesktopView) {
-      const notDetectedWallets = [...evmNotDetected, ...svmNotDetected].sort(
-        walletComparator,
-      );
-      installedWallets.push(...notDetectedWallets);
-    }
-
-    return installedWallets;
-  }, [
-    account.connector?.id,
-    chains?.types,
-    connectors,
-    isDesktopView,
-    solanaWallets,
-    walletConfig?.coinbase,
-    walletConfig?.metaMask,
-    walletConfig?.walletConnect,
-  ]);
+  const wallets = useWallets(walletConfig, chains);
 
   return (
     <PageContainer disableGutters>
